renamed package so as not to confuse console pomodoro with the new pomodoro
