+++ src/main/java/services/consolepomodororunning/CancelTimerInput.java
@@ -1,4 +1,4 @@
-package services.pomodororunning;
+package services.consolepomodororunning;
 
 import java.util.Objects;
 import java.util.Scanner;
